Props refactored
