+++ src/components/app/burger-ingredients/burger-ingrediens.module.css
@@ -9,9 +9,7 @@
 }
 
 
-.containerSub{
 
-}
 .scrollContainer{
 
     overflow-y: scroll;
