+++ src/components/app/modal-overlay/modal-overlay.js
@@ -1,5 +1,5 @@
 
-import React, {useEffect, useState} from "react";
+import React from "react";
 import styles from './modal-overlay.module.css';
 
 const ModalOverlay = () => {
